+++ solc/CommandLineInterface.cpp
@@ -122,6 +122,7 @@ static string const g_strBinaryRuntime = "bin-runtime";
 static string const g_strCombinedJson = "combined-json";
 static string const g_strCompactJSON = "compact-format";
 static string const g_strContracts = "contracts";
+static string const g_strErrorRecovery = "error-recovery";
 static string const g_strEVM = "evm";
 static string const g_strEVM15 = "evm15";
 static string const g_strEVMVersion = "evm-version";
@@ -176,6 +177,7 @@ static string const g_argCompactJSON = g_strCompactJSON;
 static string const g_argAstBoogie = g_strAstBoogie;
 static string const g_argAstBoogieArith = g_strAstBoogieArith;
 static string const g_argAstBoogieNoModAnalysis = g_strAstBoogieNoModAnalysis;
+static string const g_argErrorRecovery = g_strErrorRecovery;
 static string const g_argGas = g_strGas;
 static string const g_argHelp = g_strHelp;
 static string const g_argInputFile = g_strInputFile;
@@ -711,6 +713,7 @@ Allowed options)",
 		(g_argColor.c_str(), "Force colored output.")
 		(g_argNoColor.c_str(), "Explicitly disable colored output, disabling terminal auto-detection.")
 		(g_argNewReporter.c_str(), "Enables new diagnostics reporter.")
+		(g_argErrorRecovery.c_str(), "Enables additional parser error recovery.")
 		(g_argIgnoreMissingFiles.c_str(), "Ignore missing files.");
 	po::options_description outputComponents("Output Components");
 	outputComponents.add_options()
@@ -952,6 +955,8 @@ bool CommandLineInterface::processInput()
 		m_compiler->setSources(m_sourceCodes);
 		if (m_args.count(g_argLibraries))
 			m_compiler->setLibraries(m_libraries);
+		if (m_args.count(g_argErrorRecovery))
+			m_compiler->setParserErrorRecovery(true);
 		m_compiler->setEVMVersion(m_evmVersion);
 		// TODO: Perhaps we should not compile unless requested
 
@@ -1011,6 +1016,13 @@ bool CommandLineInterface::processInput()
 	catch (Exception const& _exception)
 	{
 		serr() << "Exception during compilation: " << boost::diagnostic_information(_exception) << endl;
+		return false;
+	}
+	catch (std::exception const& _e)
+	{
+		serr() << "Unknown exception during compilation" << (
+			_e.what() ? ": " + string(_e.what()) : "."
+		) << endl;
 		return false;
 	}
 	catch (...)
@@ -1393,6 +1405,14 @@ bool CommandLineInterface::assemble(
 			serr() << "Exception in assembler: " << boost::diagnostic_information(_exception) << endl;
 			return false;
 		}
+		catch (std::exception const& _e)
+		{
+			serr() <<
+				"Unknown exception during compilation" <<
+				(_e.what() ? ": " + string(_e.what()) : ".") <<
+				endl;
+			return false;
+		}
 		catch (...)
 		{
 			serr() << "Unknown exception in assembler." << endl;
@@ -1441,6 +1461,13 @@ bool CommandLineInterface::assemble(
 		catch (Exception const& _exception)
 		{
 			serr() << "Exception while assembling: " << boost::diagnostic_information(_exception) << endl;
+			return false;
+		}
+		catch (std::exception const& _e)
+		{
+			serr() << "Unknown exception during compilation" << (
+				_e.what() ? ": " + string(_e.what()) : "."
+			) << endl;
 			return false;
 		}
 		catch (...)