+++ Changelog.md
@@ -1,11 +1,17 @@
 ### 0.5.10 (unreleased)
 
+Important Bugfixes:
+ * Fix incorrect abi encoding of storage array of data type that occupy multiple storage slots
+
+
 Language Features:
 
 
 
 Compiler Features:
  * Optimizer: Add rule to simplify SUB(~0, X) to NOT(X).
+ * Commandline Interface: Experimental parser error recovery via the ``--error-recovery`` commandline switch.
+ * Yul Optimizer: Make the optimizer work for all dialects of Yul including eWasm.
 
 
 
@@ -44,7 +50,6 @@ Compiler Features:
  * Yul Optimizer: Optimize representation of numbers.
  * Yul Optimizer: Do not inline recursive functions.
  * Yul Optimizer: Do not remove instructions that affect ``msize()`` if ``msize()`` is used.
-
 
 Bugfixes:
  * Code Generator: Explicitly turn uninitialized internal function pointers into invalid functions when loaded from storage.
@@ -85,6 +90,7 @@ Compiler Features:
  * Yul: Adds break and continue keywords to for-loop syntax.
  * Yul: Support ``.`` as part of identifiers.
  * Yul Optimizer: Adds steps for detecting and removing of dead code.
+ * Yul Code Generator: Directly jump over a series of function definitions (instead of jumping over each one)
 
 
 Bugfixes: